Added cage trap gradual buildup

+++ src/main/java/net/geforcemods/securitycraft/items/UniversalBlockRemoverItem.java
@@ -11,9 +11,9 @@ import net.geforcemods.securitycraft.api.IOwnable;
 import net.geforcemods.securitycraft.api.LinkableBlockEntity;
 import net.geforcemods.securitycraft.api.OwnableBlockEntity;
 import net.geforcemods.securitycraft.api.Owner;
+import net.geforcemods.securitycraft.blockentities.CageTrapBlockEntity;
 import net.geforcemods.securitycraft.blockentities.DisplayCaseBlockEntity;
 import net.geforcemods.securitycraft.blockentities.InventoryScannerBlockEntity;
-import net.geforcemods.securitycraft.blocks.CageTrapBlock;
 import net.geforcemods.securitycraft.blocks.InventoryScannerBlock;
 import net.geforcemods.securitycraft.blocks.LaserBlock;
 import net.geforcemods.securitycraft.blocks.OwnableBlock;
@@ -88,8 +88,8 @@ public class UniversalBlockRemoverItem extends Item {
 					}
 				}
 				else if (block == SCContent.CAGE_TRAP.get()) {
-					if (!level.isClientSide) {
-						CageTrapBlock.disassembleIronBars(state, level, pos, owner);
+					if (!level.isClientSide() && be instanceof CageTrapBlockEntity cageTrap) {
+						cageTrap.disassembleIronBars();
 						level.destroyBlock(pos, true);
 						stack.hurtAndBreak(1, player, p -> p.broadcastBreakEvent(ctx.getHand()));
 					}
